Add close support
Currently the only way to "close" the socket was a CancellationToken
passed in the contructor. However, this did only abort the running tasks
and did not properly close the uncerlying connection.
Additionally a call to Dispose() is not guaranteed and should not be
responsible for connection closing.
The DisconnectAsync() closes the connection regardless of GC and aborts
all currently running tasks.